test/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.1
langchain-experimental>=0.0.1
neo4j>=5.0.0
streamlit>=1.0.0
pypdf>=3.0.0
python-dotenv>=1.0.0
ollama>=0.1.0
sentence-transformers>=2.0.0
unstructured>=0.10.0
pdf2image>=1.0.0